Stat collection done, going to implement stat print

+++ global.cpp
@@ -4,7 +4,7 @@ void exitNow(unsigned code){
 	exit(code);
 }
 
-cacheAtom cacheAll(const unsigned long long int& key, cacheAtom new_value)
+cacheAtom cacheAll(const uint64_t & key, cacheAtom new_value)
 {
 	return new_value;
 }
